third_party/WebKit/LayoutTests/external/wpt/cssom/getComputedStyle-pseudo-expected.txt

This is a testharness.js-based test.
PASS Resolution of width is correct for ::before and ::after pseudo-elements
PASS Resolution of width is correct for ::before and ::after pseudo-elements of display: contents elements
PASS Resolution of nonexistent pseudo-element styles
PASS Resolution of pseudo-element styles in display: none elements
PASS Item-based blockification of pseudo-elements
FAIL Item-based blockification of nonexistent pseudo-elements assert_equals: Pseudo-styles of display: flex elements should get blockified expected "block" but got "inline"
Harness: the test ran to completion.
